// src/app/Models/PersonalProposal/projection-info.ts
export class CostInfo {
    ID: number;
    CLIENT_ID: number;
    PROPOSAL_ID: number;
    COST_OF_PROJECT: string;
    COST_RUPEES: number;
}

export class MeansInfo {
    ID: number;
    CLIENT_ID: number;
    PROPOSAL_ID: number;
    MEANS_OF_FINANCE: string;
    MEANS_RUPEES: number;
}

export class ProjectionInfo {
    ID: number;
    CLIENT_ID: number;
    PROPOSAL_ID: number;
    NOTE_TO_ACHIVE_THE_PROJECTIONS:string;
}